fix/Mypage-Connect-All/2024-11-18
