+++ src/main/java/com/ohgiraffers/mergyping/config/SecurityConfig.java
@@ -37,7 +37,7 @@ public class SecurityConfig  {
         http.authorizeHttpRequests( auth -> {
             auth.requestMatchers( "/auth/**",
                     "/", "/main", "/css/**", "/img/**", "/error/**",
-                    "/userinfo/**", "/useractive/**", "/intro", "/notice/*", "/selectnotice", "/selectpost","/writepost", "/find/**",
+                    "/userinfo/**", "/useractive/**", "/intro/*", "/notice/*", "/selectnotice", "/selectpost","/writepost", "/find/**",
                     "/checknickname/**", "/admin/**", "/toggleFavorite", "/post/**","selectpost/**","/toggleScary/**","/toggleNotScary/**").permitAll();
             auth.requestMatchers("/admin/*").hasAnyAuthority(UserRole.ADMIN.getRole());
             auth.requestMatchers("/user/*").hasAnyAuthority(UserRole.USER.getRole());
