+++ src/main/java/com/ohgiraffers/mergyping/post/controller/PostController.java
@@ -59,11 +59,53 @@ public class PostController {
     @GetMapping("/post")
     public String postList(Model model) {
 
-        // 서비스를 통해 PostDTO에 있는 게시글 목록 가져옴
-        List<PostDTO> postList = postService.postList();
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
 
-        // 모델에 postList라는 이름으로 게시글 목롣 추가
-        model.addAttribute("postList", postList);
+        if (authentication != null && authentication.getPrincipal() instanceof AuthDetails) {
+            AuthDetails userDetails = (AuthDetails) authentication.getPrincipal();
+            int userNo = userDetails.getUserNo();
+
+            // 서비스를 통해 PostDTO에 있는 게시글 목록 가져옴
+            List<PostDTO> postList = postService.postList();
+
+            // 모델에 postList라는 이름으로 게시글 목롣 추가
+            model.addAttribute("postList", postList);
+
+            // 누적된 출석 수 가져오기
+            Integer attendanceCount = myPageService.getUserAttendanceCount(userNo);
+            model.addAttribute("attendanceCount", attendanceCount);
+
+            // 등급 기준 정해주기
+            int levelNo = myPageService.calculateLevel(attendanceCount);
+            System.out.println("levelNo = " + levelNo);
+
+            // 등급 기준과 출석수 기반으로 등급 업데이트하기
+            myPageService.updateUserLevel(userNo, levelNo);
+
+            // 유저의 등급 가져오기
+            String levelName = myPageService.getLevelName(levelNo);
+            model.addAttribute("userLevel", levelName);
+
+            // 유저의 다음 레벨 이름 가져오기
+            String nextLevelName = myPageService.getNextLevelName(levelNo);
+            model.addAttribute("nextLevelName", nextLevelName);
+
+            // 다음 등급에 필요한 출석 횟수 조회
+            int nextLevelRequiredAttendance = myPageService.getNextLevelRequiredAttendance(levelNo, attendanceCount);
+            model.addAttribute("nextLevelRequiredAttendance", nextLevelRequiredAttendance);
+
+
+            // MyPageDTO에 userNo를 전달하여 사용자 정보를 가져옵니다.
+            MyPageDTO myPageDTO = myPageService.findNickName(userNo);
+            model.addAttribute("myPageDTO", myPageDTO);
+
+
+            MyPageDTO userInfo = myPageService.findUserInfo(userNo);
+            model.addAttribute("userInfo", userInfo);
+
+            Map<String, Object> mbtiInfo = myPageService.findUserMBTIInfo(userNo);
+            model.addAttribute("mbtiInfo", mbtiInfo);
+        }
 
         //뷰 반환
         return "/post/post";
@@ -159,6 +201,41 @@ public class PostController {
             userNo = userDetails.getUserNo();
 
         }
+
+        // 누적된 출석 수 가져오기
+        Integer attendanceCount = myPageService.getUserAttendanceCount(userNo);
+        model.addAttribute("attendanceCount", attendanceCount);
+
+        // 등급 기준 정해주기
+        int levelNo = myPageService.calculateLevel(attendanceCount);
+        System.out.println("levelNo = " + levelNo);
+
+        // 등급 기준과 출석수 기반으로 등급 업데이트하기
+        myPageService.updateUserLevel(userNo, levelNo);
+
+        // 유저의 등급 가져오기
+        String levelName = myPageService.getLevelName(levelNo);
+        model.addAttribute("userLevel", levelName);
+
+        // 유저의 다음 레벨 이름 가져오기
+        String nextLevelName = myPageService.getNextLevelName(levelNo);
+        model.addAttribute("nextLevelName", nextLevelName);
+
+        // 다음 등급에 필요한 출석 횟수 조회
+        int nextLevelRequiredAttendance = myPageService.getNextLevelRequiredAttendance(levelNo, attendanceCount);
+        model.addAttribute("nextLevelRequiredAttendance", nextLevelRequiredAttendance);
+
+
+        // MyPageDTO에 userNo를 전달하여 사용자 정보를 가져옵니다.
+        MyPageDTO myPageDTO = myPageService.findNickName(userNo);
+        model.addAttribute("myPageDTO", myPageDTO);
+
+
+        MyPageDTO userInfo2 = myPageService.findUserInfo(userNo);
+        model.addAttribute("userInfo", userInfo2);
+
+        Map<String, Object> mbtiInfo = myPageService.findUserMBTIInfo(userNo);
+        model.addAttribute("mbtiInfo", mbtiInfo);
 
         // 서비스를 통해 게시글 번호로 게시물 조회
         SelectPostDTO selected = postService.selectById(postNo);
@@ -526,6 +603,41 @@ public class PostController {
             AuthDetails userDetails = (AuthDetails) authentication.getPrincipal();
             int userNo = userDetails.getUserNo();
             System.out.println("userNo = " + userNo);
+
+            // 누적된 출석 수 가져오기
+            Integer attendanceCount = myPageService.getUserAttendanceCount(userNo);
+            model.addAttribute("attendanceCount", attendanceCount);
+
+            // 등급 기준 정해주기
+            int levelNo = myPageService.calculateLevel(attendanceCount);
+            System.out.println("levelNo = " + levelNo);
+
+            // 등급 기준과 출석수 기반으로 등급 업데이트하기
+            myPageService.updateUserLevel(userNo, levelNo);
+
+            // 유저의 등급 가져오기
+            String levelName = myPageService.getLevelName(levelNo);
+            model.addAttribute("userLevel", levelName);
+
+            // 유저의 다음 레벨 이름 가져오기
+            String nextLevelName = myPageService.getNextLevelName(levelNo);
+            model.addAttribute("nextLevelName", nextLevelName);
+
+            // 다음 등급에 필요한 출석 횟수 조회
+            int nextLevelRequiredAttendance = myPageService.getNextLevelRequiredAttendance(levelNo, attendanceCount);
+            model.addAttribute("nextLevelRequiredAttendance", nextLevelRequiredAttendance);
+
+
+            // MyPageDTO에 userNo를 전달하여 사용자 정보를 가져옵니다.
+            MyPageDTO myPageDTO = myPageService.findNickName(userNo);
+            model.addAttribute("myPageDTO", myPageDTO);
+
+
+            MyPageDTO userInfo2 = myPageService.findUserInfo(userNo);
+            model.addAttribute("userInfo", userInfo2);
+
+            Map<String, Object> mbtiInfo = myPageService.findUserMBTIInfo(userNo);
+            model.addAttribute("mbtiInfo", mbtiInfo);
 
             MyPageDTO userInfo = myPageService.findUserInfo(userNo);
             model.addAttribute("userInfo", userInfo);
